Tests now actually fully run under Firefox 9.

+++ test/test.js
@@ -1,4 +1,6 @@
-window.sessionStorage && sessionStorage.clear();
+if (! $.browser.mozilla ) {
+	window.sessionStorage && sessionStorage.clear();
+}
 jQuery.validator.defaults.debug = true;
 $.mockjaxSettings.log = $.noop;
